🔨Merge pull request (#96) to refactor (in very future) utilities/convert or make it easier to maintain.
🚧  add @abstractmethod on method in base class CoordinateConverter
🚧  Remove transform_array which has not been used in other functions
🚧  Add FuncAnimation in type hint for as_moview
🚨  add ignore reportUnusedImport in __init__.py
🐛  use cycle instead of "ch1" in prodigy_itx.py

* 🔨  Tiny refactoring codes

🏁  Add warning in find_first_file

✅  Add test for convert_itx_format
✅  Update test for convert_itx_format and revise methods in ProdigyItx
⚡️  Update delaytime_fs to accept units argument.
🔧  Remove A005 from ruff check.

Add @abstractmethod in conversion/base.py

⚡️  Cherry pick from 08a9553c to fix issue #97.

+++ src/arpes/bootstrap.py
@@ -238,7 +238,7 @@ class Normal(Distribution):
     def draw_samples(
         self,
         n_samples: int = Distribution.DEFAULT_N_SAMPLES,
-    ) -> NDArray[np.int_]:
+    ) -> NDArray[np.float64]:
         """Draws samples from this distribution."""
         return scipy.stats.norm.rvs(
             self.center,
